Protect todos with JWT middleware

+++ todo-backend/index.js
@@ -3,6 +3,7 @@ import dotenv from "dotenv"
 import cors from "cors";
 import pool from "./db.js";
 import authRoutes from "./routes/auth.js";
+import authMiddleware from "./middleware/authMiddleware.js";
 
 
 dotenv.config();
@@ -19,9 +20,9 @@ app.use("/auth", authRoutes);
 
 
 
-app.get("/todos", async (req,res) => {
+app.get("/todos", authMiddleware, async (req,res) => {
 try {
-    const result = await pool.query("SELECT * FROM todos");
+    const result = await pool.query("SELECT * FROM todos WHERE user_id = $1",[req.user.userID]);
     res.json(result.rows);
 } catch (err) {
     console.error(err);
@@ -34,10 +35,10 @@ app.get("/", (req, res) => {
     res.send("the Server is Running");
 })
 
-app.post("/todos", async (req,res) => {
+app.post("/todos", authMiddleware, async (req,res) => {
 try {
     const {text} = req.body;
-    const result = await pool.query("INSERT INTO todos (text) VALUES ($1) RETURNING *",[text]);
+    const result = await pool.query("INSERT INTO todos (text, user_id) VALUES ($1, $2) RETURNING *",[text, req.user.userID]);
     res.json(result.rows[0]);
 } catch (err) {
     console.error(err);
@@ -45,17 +46,17 @@ try {
 }
 })
 
-app.delete("/todos/:id", async (req, res) => {
+app.delete("/todos/:id", authMiddleware, async (req, res) => {
     const id = Number(req.params.id);
     if(isNaN(id)) {
         return res.status(400).json({error: "Invalid ID"})
     }
 
 try {
-    const result = await pool.query("DELETE FROM todos WHERE id = $1 RETURNING *", [id]);
+    const result = await pool.query("DELETE FROM todos WHERE id = $1 AND user_id =$2 RETURNING *", [id, req.user.userID]);
 
     if(result.rowCount === 0) {
-        return res.status(400).json({error: "Todo Not Found"});
+        return res.status(400).json({error: "Not Authorized"});
     }
 
     res.json(result.rows[0]);
@@ -66,7 +67,7 @@ try {
 }
 })
 
-app.put("/todos/:id", async (req, res) => {
+app.put("/todos/:id", authMiddleware, async (req, res) => {
     const id = Number(req.params.id);
     const updatedText = req.body.text
 
@@ -75,7 +76,7 @@ app.put("/todos/:id", async (req, res) => {
     }
 
     try {
-        const result = await pool.query(`UPDATE todos SET text = $1 WHERE id = $2 RETURNING *`,[updatedText,id]);
+        const result = await pool.query(`UPDATE todos SET text = $1 WHERE id = $2 AND user_id =$3 RETURNING *`,[updatedText,id, req.user.userID]);
         res.json(result.rows[0])
 
     } catch (err) {
